refactor: rename middleware to proxy
