+++ src/components/SectionEntry.tsx
@@ -2,7 +2,7 @@
 
 import { ReactNode, useState } from 'react';
 import { ChevronDown } from 'lucide-react';
-import { cn } from '@/src/lib/utils';
+import { cn } from '@/lib/utils';
 
 interface SectionEntryProps {
   title: ReactNode;
